refactor: move prompt related browse components to prompts folder
