+++ app/prompts/my/page.tsx
@@ -1,6 +1,6 @@
 import { searchPrompts } from "@/app/lib/actions/prompts";
 import SearchBox from "@/app/ui/common/search";
-import SearchResults from "@/app/ui/browse/search-result";
+import SearchResults from "@/app/ui/prompts/browse/search-result";
 import CreatePromptButton from "@/app/ui/prompts/create-prompt-button";
 import { Suspense } from "react";
