+++ app/ui/prompts/browse/search-result.tsx
